requirements.txt
fastapi
uvicorn
openai
python-dotenv
markitdown[all] 
python-multipart 
pypdf>=4.2.0